Added SNR calculations per signal, which resulted in changing the outputs of the quickSignalSingle and how it interacts with the rest of the sim.
I was having issues with seeding not causing reproducable signals etc. when running with noise v.s.
without.  This was originally solved by just saving the state before adding noise and resetting it after
such that the random state is unaffected by the noise being called.  This was not a great work around
as it meant all signals had the same noise unless something else called random.  The thing that would
change the state was the inelasticity calculation.  So each event had the same noise across each antenna
(though different from each event, as this was calculated per event).  Not I have taking out the state
resetting and have chosen to do all randomness that shouldn't be affected by noise upfront.  I.e. I have
moved the inelasticity calculation to be earlier and for all events at once.  This way it is calculated
before any additional random calls have a chance to alter the queue.  This is something to be careful
of if future additions depend on numpy.random.

I have added some table creation for event viewing in the askaryan_testing.py file and plan on adding
that to the actual saved plots from the simualtion.

Additional future plans:
- Add beam patterns to each antenna.
- Add ideal beam-formed signal to output plots
- Include SNR in output/saved simulation data (currently calculated and unused)
- Add histogram of SNR (probably after sim in an analysis script)
- Improve attenuation to be frequency dependant.  This will likely be very tricky as it is calculated
  during the ray tracing library, so doing it per frequency is crazy.
- Speed up loading of ray tracing libraries and use of griddata on these libraries by taking advantage
  of multiprocessing.  Each of these interpolations are independant for each antenna and could be done
  in parallel when multiple cores are available.
- Large interpolation speed increases could be obtained if I find a way to pass griddata the delaunay
  triangle locations.  Currently for each variable it interpolates it needs to locate where the events
  are on the grid.  This location doesn't change for each variable needing to be interpolated and
  requires a lot of computation time.
  A possible fix would be to do what Cosmin keeps suggesting which is to make a rectalinear lookup table
  or something.  I am in no way eager to look into this and this is on the back burner.
- Eventually add the beamforming like done in the real array.   I talked to Eric about this and will
  probably try and tackle it once everything else seems in order.