Fully added in beam pattern support.  This is currently only for simple (1 everywhere) or dipole beam patterns.  Orientation of antennas has been added, which will allow for implementations of new beam patterns easier, as they can be defined in the coordinate system of the antenna easily in the same way the dipole calculation is done.
I also added a function to askaryan.py that digitizes a signal to a given bit range, and samples at
a given sampling rate (with slight time jitter support to avoid biasing results by having constant
alignment between signal 0 second point and antenna 0 second point).  There is also an option for
adding in a DC offset before the digitization works so that the signal can be rounded with that
included.

In the next commit I should have this digitization implemented into the simulation.  I am currently not
100% clear how I want the out of this to interact with the rest of the code.  It is likely that
triggering will be done on this digitized signal (perhaps with beam forming, or a simple threshold).
I recommend to future Dan to add a uniform time offset (calculated in advance in 1 array) of time
offsets from -1ns to 1ns, and an array (potentially all of the same value, but an array incase
differences are desired later) of dc_offset terms.  The scale_noise_from should be calculated in advance
using a signal with attenuation = 0 such that only noise is returned.  std of this describes noise level.
scale_noise_to should be 3.

The digitization might also make the SNR output not right, as it is calculated before the signal is
capped (which would in theory cap off the SNR at 64/3 (where noise should be set to 3). I will need to
think about this as well.