Mostly added a lot to the beamforming_testing.py file.  It currently has functions to sync signals temporally, to calculate/produce a dictionary containing all of the info about the different possible beams for an antenna array, and a function that uses this dict with the synced signal inputs to do the beam summing.
I intend to next add in the power summing algorithm to these functions in the testing
stage before adding all of this to the main simulation.

Additionally in the future I wish to fix a problem currently identified with summed
signals (two signals added in the time domain which are close enough temporally that
their noise profiles overlap).  Overlapping noise adds and artificially increases the
peaks and troughs of the noise in this region.  I think they way that I will fix this
is by subtracting out the noise in the regions where the signals overlap, or by doing
some ramping addition of noise from each signal.   Because of how I am doing the SNR
calculation I shoud have both the raw signal, and the signal with the noise added,
so I can just subtract the raw signal from the noisy signal to just get the noise
profile.