Updated the function that does the beamforming and power sum to not alter the incoming dictionary, as I learned that unlike most things passed to functions, dictionaries are not internally copied and thus alters to it are saved outside of the function as well.  Using better practices here now
Additionally I fixed a bug with how the correct solution type was being saved
in the simulation.  I am not sure how long this bug has been there but before
this fix only the direct solution was output in the info object.  Now the
solution type with max magnitude is saved.  This may produce some hard to interpret
outputs when the signal is small as the noise ensures the max solution type
magnitude to be essentially random.  I am still undecided if I want to just
output the info for all solution types (I probably should).