I made a lot of changes to some organizational stuff.  Added a triggered flag to the output info array, removed the trigger times which were kind of meaningless, and fixed the max adu per solution type output.
Crated the testing_single_event.py to more easily test things offline.  It essentially
emulates the main event function in antarcticsim (time domain only) given an reader input
and event id.  This is pretty helpful for producing the plots after the fact as well and
getting better access to say, beam forming.  Note that as with ay other 'offline' function
I have made like this, the noise will not match the same as that in the actual simulation
as they will be at different points in the seed.  But for most purposes that is fine.

Tested the newest versions of Dan Smith's antenna and system responses however we are not
quite satisfied with how things look, so some work is still under way.