Made some changes while producing VA plots etc.  Changed plotGeometry to take the config file, which is not less prone to mistake (which was happening before).
I have noticed that the current state of the code (which historically derives
from a greenland simualtion, which was then ported to an antartic version)
exists as two unequally maintained simulations.  My development has consistently
been on antarcticsim and functions called by it.  I have noticed that some files
that are called that haven't pertained closely to my work were not fully transformed
for the antarctic simulation.  Meaning in functions used by antarcticsim such as
prem.prem used for the density function of the Earth ALWAYS use the greenland ice
model.

I hope in the next commit to have fixed this, changing the code from two seperate
simulations to one, with a simple flag to declare the region you want the simulation
to operate.