I completely reworked a lot of the organization of the simulation.  The station and antenna classes were previously extremely underutilized.  Many additional dictionaries were being used to store information that is not just attached to the antenna or station objects.
These changes had very far reaching effects, as they are essential components of the simulation,
and many functons that were previously externally defined are not defined as methods within
these classes.  Thus even a lot of the analysis scripts had to be changed quite a bit to
account for this.

Overall I think it is far better, however as so many changes were made I may have not
fully accounted for them, and certain untested paths may break if used in the future.
Hopefully I will be the one to find these and fix them.

This reorganzation should make adding polarization support more sensible, as the antenna
framework has been built out in anticipation of it.