Added a first pass at polarization and antenna's response to that polarization. Hopefully this was doe correctly.  Currently have tested it with small simulations to see if anything broke, which does not seem to be the case.  Signals still look reasonable.  Will be running larger tests to see how volumetric acceptance has changed.
Changed from using Fresnel power coefficients to Fresnel amplitude, such that the sign can be obtained
for certain reflections.  Hopefully this doesn't introduce issues.  I am only taking the real component
of this, so if there is a need for dielectrics at some later point, or if something seems weird, that
could be a thing to check.  For the current libraries it is not an issue.

Also changes a_h -> a_s and a_v -> a_p, as this is more accurately what they represent.