Significantly updated the README file.
Added some docstrings.

Altered volumetric_acceptance.py to work with both merged and non-merged simultaneously.

Renamed quat to linalg, as it has become more of a generic linear algebra module.

Changed the polarization calculations and beam pattern calculations to expect the vectors to be pre-calculated.

Enforced a phi, theta convention for input parameters.

Took away some uses of normalize when the units involved should already be normalized.

Fixed the calculation for the polarization, as it was defined incorrectly.  Double checked this calculation.

Used more explicit names for some of the vectors used throughout.

Added the projection.py file to utils.  This calculates the projected visible area of a circle on the Earths surface
as viewed from different angles.  This is something I made for Cosmin for ANITA stuff, but might be useful here
some day so I have included it.