Enforced a convention for input and output voltages and times where I could. Moved plotting function into the Station class.
Redeveloped how station configuration files should be, now allowing for distinction
between reconstruction antennas and phased array antennas.

Reorganized how event works to account for new distinction between antenna types.
Triggers are now only attempted on phased array antennas, and if the trigger is not
passed by these antennas then the reconstruction antennas are not calculated.

Reconstruction antennas do count in the pre trigger however, so the more reconstruction
antennas the more likely a pre trigger will be met.

Beam forming preparations and calculations will only be operated on the phased array antennas.
These antennas should all be located on a single vertical string.

Updated the README corresponding to changes (may have missed some).

Updated plotGeometry to plot in 3d as an option, which becomes more helpful for more complicated
station configurations.  I also added support for plotting neutrino and polarization vectors as well.

Moved the profiling function to gnosim/utils/misc

Because one station can trigger without another the trigger bool is not necessarily the same
for each entry in info for an event.  I added some code to the volumetric acceptance calculation
that should make pull ou the desired effect for that calculation of True if any station sees the event
and False if none do.  Before because triggering was event wide I could just do a simple unique call
but now I had to add a couple of lines.  This hasn't really been tested yet.

Threshold units and trigger level were defined in both configuration file types so I took them out of
the detector config file.  In the future we may want to be able to apply a different trigger for
each station but that is currently not the case.

Included an absolute value for triggering when using 'V' or 'adu' triggering.

Adjusted the info_dtype to use appropriate int types for some of the numbers.

Fixed the observation angles, which were off due to one vector being incorrectly used.

I am currently still double checking if the change volumetric acceptances due to the addition
of polarizaiton makes sense. So far it looks correct.